internal/ui: switch back to manual GtkBuilder usage
Maybe rewriting in C will make sense at some point, but not quite yet.
