+++ internal/ui/preferences.go
@@ -4,13 +4,11 @@ import (
 	_ "embed"
 
 	"github.com/diamondburned/gotk4-adwaita/pkg/adw"
-	coreglib "github.com/diamondburned/gotk4/pkg/core/glib"
-	"github.com/diamondburned/gotk4/pkg/glib/v2"
 	"github.com/diamondburned/gotk4/pkg/gtk/v4"
 )
 
 //go:embed preferences.ui
-var preferencesXML []byte
+var preferencesXML string
 
 type PreferencesWindow struct {
 	adw.PreferencesWindow
@@ -19,14 +17,6 @@ type PreferencesWindow struct {
 	UseTrayIcon    *gtk.Switch
 }
 
-var preferencesWindowType = coreglib.RegisterSubclass[*PreferencesWindow](
-	coreglib.WithClassInit(func(class *gtk.WidgetClass) {
-		class.SetTemplate(glib.NewBytesWithGo(preferencesXML))
-	}),
-)
-
 func NewPreferencesWindow() *PreferencesWindow {
-	win := preferencesWindowType.New()
-	win.InitTemplate()
-	return win
+	return newFromBuilder[PreferencesWindow](preferencesXML)
 }
