+++ internal/ui/trayscale.cmb
@@ -2,9 +2,9 @@
 <!DOCTYPE cambalache-project SYSTEM "cambalache-project.dtd">
 <cambalache-project version="0.11.3" target_tk="gtk-4.0">
   <ui>
-	(1,1,"mainwindow.ui","mainwindow.ui",None,None,None,None,None,None,None),
-	(2,3,None,"peerpage.ui",None,None,None,None,None,None,None),
-	(3,1,None,"preferences.ui",None,None,None,None,None,None,None)
+	(1,None,"mainwindow.ui","mainwindow.ui",None,None,None,None,None,None,None),
+	(2,None,None,"peerpage.ui",None,None,None,None,None,None,None),
+	(3,None,None,"preferences.ui",None,None,None,None,None,None,None)
   </ui>
   <object>
 	(1,1,"AdwApplicationWindow","MainWindow",None,None,None,None,None,None),
