--- internal/ui/ui.c
@@ -1,6 +1,0 @@
-#include <gtk/gtk.h>
-
-#include "ui.h"
-
-void _run_ui(void) {
-}
