--- internal/ui/ui.h
@@ -1,1 +1,0 @@
-void _run_ui(void);
